+++ backend/src/playback/dto/current-playback.dto.ts
@@ -1,27 +1,8 @@
 import { ApiProperty } from "@nestjs/swagger";
 import { Expose, Type } from "class-transformer";
 
-export class CurrentPlaybackDeviceDto {
-  @ApiProperty()
-  @Expose()
-  id: string;
-
-  @ApiProperty()
-  @Expose()
-  isActive: boolean;
-
-  @ApiProperty()
-  @Expose()
-  name: string;
-
-  @ApiProperty()
-  @Expose()
-  type: string;
-
-  @ApiProperty()
-  @Expose()
-  volumePercent: number;
-}
+import { CurrentPlaybackDeviceDto } from "./current-playback-device.dto";
+import { CurrentPlaybackTrackDto } from "./current-playback-track.dto";
 
 export class CurrentPlaybackStateDto {
   @ApiProperty({ nullable: true, type: CurrentPlaybackDeviceDto })
@@ -49,52 +30,4 @@ export class CurrentPlaybackStateDto {
   @Expose()
   @Type(() => CurrentPlaybackTrackDto)
   track: CurrentPlaybackTrackDto | null;
-}
-
-export class CurrentPlaybackTrackAlbumDto {
-  @ApiProperty()
-  @Expose()
-  id: string;
-
-  @ApiProperty({ type: [String] })
-  @Expose()
-  images: string[];
-
-  @ApiProperty()
-  @Expose()
-  name: string;
-}
-
-export class CurrentPlaybackTrackArtistDto {
-  @ApiProperty()
-  @Expose()
-  id: string;
-
-  @ApiProperty()
-  @Expose()
-  name: string;
-}
-
-export class CurrentPlaybackTrackDto {
-  @ApiProperty({ type: CurrentPlaybackTrackAlbumDto })
-  @Expose()
-  @Type(() => CurrentPlaybackTrackAlbumDto)
-  album: CurrentPlaybackTrackAlbumDto;
-
-  @ApiProperty({ type: [CurrentPlaybackTrackArtistDto] })
-  @Expose()
-  @Type(() => CurrentPlaybackTrackArtistDto)
-  artists: CurrentPlaybackTrackArtistDto[];
-
-  @ApiProperty()
-  @Expose()
-  durationMs: number;
-
-  @ApiProperty()
-  @Expose()
-  id: string;
-
-  @ApiProperty()
-  @Expose()
-  name: string;
 }
